pyproject.toml
[project]
name = "nonebot_plugin_osubot"
version = "6.26.4"
description = "OSUbot in NoneBot2"
authors = [
    {name = "yaowan233", email = "572473053@qq.com"},
]
license = {text = "AGPL-3.0"}
dependencies = [
    "nonebot2>=2.3.0",
    "pydantic>=1.10.0,<3.0.0,!=2.5.0,!=2.5.1",
    "nonebot-plugin-alconna>=0.46.4",
    "nonebot-plugin-session~=0.3",
    "pillow>=9.2.0",
    "expiringdict>=1.2.2",
    "nonebot-plugin-apscheduler>=0.4.0",
    "tortoise-orm>=0.20.0",
    "nonebot-plugin-tortoise-orm>=0.1.4",
    "rosu-pp-py==3.1.0",
    "reamber>=0.2.1",
    "httpx>=0.23.3",
    "typing_extensions>=4.11.0",
    "matplotlib>=3.7.1",
    "nonebot-plugin-htmlrender>=0.3.1",
    "nonebot-plugin-waiter>=0.6.1",
    "nonebot-adapter-onebot>=2.4.6",
]
requires-python = ">=3.9.8,<3.13"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/yaowan233/nonebot-plugin-osubot"
Repository = "https://github.com/yaowan233/nonebot-plugin-osubot"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[tool.pdm.dev-dependencies]
test = [
    "isort>=5.13.2",
    "black>=24.4.2",
    "loguru>=0.7.2",
    "ruff>=0.4.2",
    "nonemoji>=0.1.4",
    "nonebug>=0.3.5",
    "pytest-asyncio>=0.21.1",
    "fastapi>=0.103.1",
    "uvicorn>=0.23.2",
    "pre-commit>=3.7.1",
]

[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 120
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901", "T201", "E731", "E402"]
